supporting location specific parameters on the site and building object now, including climate zone, resolves #89
