+++ lib/buildingsync/model_articulation/site.rb
@@ -36,7 +36,7 @@
 # *******************************************************************************
 require_relative 'building'
 module BuildingSync
-  class Site < SpatialElement
+  class Site < LocationElement
     # initialize
     def initialize(build_element, ns)
       # code to initialize
@@ -46,26 +46,10 @@ module BuildingSync
       @premises_notes = nil
       @all_set = false
 
-
-      # parameter to read and write.
-      @climate_zone = nil
-      @climate_zone_ashrae = nil
-      @climate_zone_ca_t24 = nil
-      @weather_file_name = nil
-      @weather_station_id = nil
-      @city_name = nil
-      @state_name = nil
-      @latitude = nil
-      @longitude = nil
-      @street_address = nil
-      @postal_code = nil
-
-
       # TM: just use the XML snippet to search for the buildings on the site
       read_xml(build_element, ns)
     end
 
-    # adding a site to the facility
     def read_xml(build_element, ns)
       # first we check if the number of buildings is ok
       number_of_buildings = 0
@@ -83,16 +67,8 @@ module BuildingSync
       @occupancy_type = read_occupancy_type(build_element, nil, ns)
       # check floor areas at the site level
       @total_floor_area = read_floor_areas(build_element, nil, ns)
-      # read in the ASHRAE climate zone
-      read_climate_zone(build_element, ns)
-      # read in the weather station name
-      read_weather_file_name(build_element, ns)
-      # read city and state name
-      read_city_and_state_name(build_element, ns)
-      # read latitude and longitude
-      read_latitude_and_longitude(build_element, ns)
-      # read site address
-      read_site_other_details(build_element, ns)
+      # read location specific values
+      read_location_values(build_element, ns)
       # code to create a building
       build_element.elements.each("#{ns}:Buildings/#{ns}:Building") do |buildings_element|
         @buildings.push(Building.new(buildings_element, @occupancy_type, @total_floor_area, ns))
@@ -103,81 +79,6 @@ module BuildingSync
       if !@all_set
         @all_set = true
         @buildings.each(&:set_all)
-      end
-    end
-
-    def read_climate_zone(build_element, ns)
-      if build_element.elements["#{ns}:ClimateZoneType/#{ns}:ASHRAE"]
-        @climate_zone_ashrae = build_element.elements["#{ns}:ClimateZoneType/#{ns}:ASHRAE/#{ns}:ClimateZone"].text
-      else
-        @climate_zone_ashrae = nil
-      end
-      if build_element.elements["#{ns}:ClimateZoneType/#{ns}:CaliforniaTitle24"]
-        @climate_zone_ca_t24 = build_element.elements["#{ns}:ClimateZoneType/#{ns}:CaliforniaTitle24/#{ns}:ClimateZone"].text
-      else
-        @climate_zone_ca_t24 = nil
-      end
-      if @climate_zone_ca_t24.nil? && @climate_zone_ashrae.nil?
-        OpenStudio.logFree(OpenStudio::Warn, 'BuildingSync.Site.read_climate_zone', 'Could not find a climate zone in the BuildingSync file.')
-      end
-    end
-
-    def read_weather_file_name(build_element, ns)
-      if build_element.elements["#{ns}:WeatherStationName"]
-        @weather_file_name = build_element.elements["#{ns}:WeatherStationName"].text
-      else
-        @weather_file_name = nil
-      end
-      if build_element.elements["#{ns}:WeatherDataStationID"]
-        @weather_station_id = build_element.elements["#{ns}:WeatherDataStationID"].text
-      else
-        @weather_station_id = nil
-      end
-    end
-
-    def read_city_and_state_name(build_element, ns)
-      if build_element.elements["#{ns}:Address/#{ns}:City"]
-        @city_name = build_element.elements["#{ns}:Address/#{ns}:City"].text
-      else
-        @city_name = nil
-      end
-      if build_element.elements["#{ns}:Address/#{ns}:State"]
-        @state_name = build_element.elements["#{ns}:Address/#{ns}:State"].text
-      else
-        @state_name = nil
-      end
-    end
-
-    def read_site_other_details(build_element, ns)
-      if build_element.elements["#{ns}:Address/#{ns}:StreetAddressDetail/#{ns}:Simplified/#{ns}:StreetAddress"]
-        @street_address = build_element.elements["#{ns}:Address/#{ns}:StreetAddressDetail/#{ns}:Simplified/#{ns}:StreetAddress"].text
-      else
-        @street_address = nil
-      end
-
-      if build_element.elements["#{ns}:Address/#{ns}:PostalCode"]
-        @postal_code = build_element.elements["#{ns}:Address/#{ns}:PostalCode"].text.to_i
-      else
-        @postal_code = nil
-      end
-
-      if build_element.elements["#{ns}:PremisesNotes"]
-        @premises_notes = build_element.elements["#{ns}:PremisesNotes"].text
-      else
-        @premises_notes = nil
-      end
-    end
-
-    def read_latitude_and_longitude(build_element, ns)
-      if build_element.elements["#{ns}:Latitude"]
-        @latitude = build_element.elements["#{ns}:Latitude"].text
-      else
-        @latitude = nil
-      end
-      if build_element.elements["#{ns}:Longitude"]
-        @longitude = build_element.elements["#{ns}:Longitude"].text
-      else
-        @longitude = nil
       end
     end
 
@@ -236,7 +137,11 @@ module BuildingSync
     end
 
     def get_climate_zone
-      return @climate_zone
+      if @climate_zone.nil?
+        return get_largest_building.get_climate_zone
+      else
+        return @climate_zone
+      end
     end
 
     def get_building_objects
@@ -272,6 +177,8 @@ module BuildingSync
       @climate_zone = @climate_zone_ashrae
       # for now we use the california climate zone if it is available
       @climate_zone = @climate_zone_ca_t24 if !@climate_zone_ca_t24.nil? && standard_to_be_used == CA_TITLE24
+      @climate_zone = building.get_climate_zone(standard_to_be_used) if @climate_zone.nil?
+      OpenStudio.logFree(OpenStudio::Warn, 'BuildingSync.Site.generate_baseline_osm', 'Could not find a climate zone in the BuildingSync file.') if @climate_zone.nil?
       building.set_weather_and_climate_zone(@climate_zone, epw_file_path, standard_to_be_used, @latitude, @longitude, ddy_file, @weather_file_name, @weather_station_id, @state_name, @city_name)
       building.generate_baseline_osm(standard_to_be_used)
     end
@@ -298,7 +205,7 @@ module BuildingSync
       site.elements["#{ns}:Latitude"].text = @latitude if !@latitude.nil?
       site.elements["#{ns}:Longitude"].text = @longitude if !@longitude.nil?
 
-      write_parameters_to_xml_for_spatial_element(ns, site)
+      write_parameters_to_xml_for_spatial_element(site, ns)
 
       site.elements.each("#{ns}:Buildings/#{ns}:Building") do |buildings_element|
         @buildings[0].write_parameters_to_xml(ns, buildings_element)
