+++ spec/files/L000_OpenStudio_Pre-Simulation_02.xml
@@ -1,0 +1,63 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<auc:BuildingSync xmlns:auc="http://buildingsync.net/schemas/bedes-auc/2019" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://buildingsync.net/schemas/bedes-auc/2019 https://raw.githubusercontent.com/BuildingSync/schema/v2.2.0/BuildingSync.xsd" version="2.2.0">
+  <auc:Facilities>
+    <auc:Facility ID="Facility1">
+      <auc:Sites>
+        <auc:Site ID="Site1">
+          <auc:Buildings>
+            <auc:Building ID="Office_Carolina">
+              <auc:PremisesName>Office Carolina</auc:PremisesName>
+              <!-- Specify either City/State, Climate Zone, or Lat/Long -->
+              <auc:ClimateZoneType>
+                <auc:ASHRAE>
+                  <auc:ClimateZone>6A</auc:ClimateZone>
+                </auc:ASHRAE>
+              </auc:ClimateZoneType>
+              <auc:BuildingClassification>Commercial</auc:BuildingClassification>
+              <auc:OccupancyClassification>Office</auc:OccupancyClassification>
+              <auc:FloorAreas>
+                <auc:FloorArea>
+                  <auc:FloorAreaType>Gross</auc:FloorAreaType>
+                  <auc:FloorAreaValue>31053</auc:FloorAreaValue>
+                </auc:FloorArea>
+              </auc:FloorAreas>
+              <auc:YearOfConstruction>1915</auc:YearOfConstruction>
+            </auc:Building>
+          </auc:Buildings>
+        </auc:Site>
+      </auc:Sites>
+      <auc:Reports>
+        <auc:Report ID="Report1">
+          <auc:Scenarios>
+            <auc:Scenario ID="Baseline-Modeled">
+              <auc:ScenarioName>Baseline</auc:ScenarioName>
+              <!--
+                Declaring the ScenarioType in this way is used by the BuildingSync-gem to create
+                the 'Baseline' model of the building.  It is intended to be interpreted as: 'Create
+                a model for the building as it currently exists'
+              -->
+              <auc:ScenarioType>
+                <auc:CurrentBuilding>
+                  <auc:CalculationMethod>
+                    <auc:Modeled>
+                      <auc:SimulationCompletionStatus>Not Started</auc:SimulationCompletionStatus>
+                    </auc:Modeled>
+                  </auc:CalculationMethod>
+                </auc:CurrentBuilding>
+              </auc:ScenarioType>
+              <!--
+                Link the Scenario back to the Building so as to specify for which building
+                the model is being created.
+              -->
+              <auc:LinkedPremises>
+                <auc:Building>
+                  <auc:LinkedBuildingID IDref="Office_Carolina"/>
+                </auc:Building>
+              </auc:LinkedPremises>
+            </auc:Scenario>
+          </auc:Scenarios>
+        </auc:Report>
+      </auc:Reports>
+    </auc:Facility>
+  </auc:Facilities>
+</auc:BuildingSync>
